backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
google-play-scraper==1.2.6
instaloader==4.10.3
python-dotenv==1.0.0
apscheduler==3.10.4
